refactor(auth-page): change the auth page name to be auth pages so that make it's children in the same directory
